+++ src/pages/CadastroMarca/index.js
@@ -1,0 +1,3 @@
+import CadastroMarca from "./CadastroMarca";
+
+export default CadastroMarca;
