+++ src/pages/CadastroMarca/styles.js
@@ -1,0 +1,9 @@
+import { makeStyles } from "@material-ui/core"
+
+const useStyles = makeStyles(() => ({
+  actions: {
+      marginLeft: '10px',
+  }
+}));
+
+export default useStyles;
